+++ code/src/app/[lang]/@modal/(.)edit/[type]/[id]/page.tsx
@@ -3,7 +3,7 @@ import { Metadata } from 'next';
 import { Modules, PageTitles } from 'io/config/routes';
 import SettingsStore from 'io/config/settings';
 import { NavBarItemSettings, UISettings } from 'types/settings';
-import FormContainerComponent from 'ui/interaction/form/form-container';
+import { InterceptFormContainerComponent } from 'ui/interaction/form/form-container';
 
 interface InterceptEditFormPageProps {
   params: Promise<{
@@ -33,11 +33,10 @@ export default async function InterceptEditFormPage(props: Readonly<InterceptEdi
   const uiSettings: UISettings = SettingsStore.getUISettings();
   const decodedType = decodeURIComponent(resolvedParams?.type);
   return (
-    <FormContainerComponent
+    <InterceptFormContainerComponent
       entityType={decodedType}
       formType={'edit'}
       isPrimaryEntity={uiSettings?.resources?.registry?.data === decodedType}
-      isModal={true}
     />
   );
 }