+++ code/src/app/[lang]/delete/[type]/[id]/page.tsx
@@ -3,7 +3,7 @@ import { Metadata } from 'next';
 import { Modules, PageTitles } from 'io/config/routes';
 import SettingsStore from 'io/config/settings';
 import { NavBarItemSettings, UISettings } from 'types/settings';
-import FormContainerComponent from 'ui/interaction/form/form-container';
+import { FormContainerComponent } from 'ui/interaction/form/form-container';
 
 interface DeleteFormPageProps {
   params: Promise<{
@@ -39,7 +39,6 @@ export default async function DeleteFormPage(props: Readonly<DeleteFormPageProps
       entityType={decodedType}
       formType={'delete'}
       isPrimaryEntity={uiSettings?.resources?.registry?.data === decodedType}
-      isModal={false}
     />
   );
 }