+++ code/src/ui/interaction/form/form-container.tsx
@@ -19,20 +19,33 @@ import { genBooleanClickHandler } from "utils/event-handler";
 import { makeInternalRegistryAPIwithParams } from "utils/internal-api-services";
 import RedirectButton from "../action/redirect/redirect-button";
 import Button from "../button";
-import Drawer from "../drawer/drawer";
 import { ENTITY_STATUS, FORM_STATES, translateFormType } from "./form-utils";
 import { FormTemplate } from "./template/form-template";
 
-import { Routes } from "io/config/routes";
-import { useSelector } from "react-redux";
-import { getCurrentEntityType } from "state/registry-slice";
 import { toast } from "../action/toast/toast";
+import NavigationDrawer from "../drawer/navigation-drawer";
 
 interface FormContainerComponentProps {
   entityType: string;
   formType: FormType;
   isPrimaryEntity?: boolean;
-  isModal?: boolean;
+}
+
+/**
+ * Renders a form container for intercept routes.
+ *
+ * @param {string} entityType The type of entity.
+ * @param {FormType} formType The type of form such as add, update, delete, and view.
+ * @param {boolean} isPrimaryEntity An optional indicator if the form is targeting a primary entity.
+ */
+export function InterceptFormContainerComponent(
+  props: Readonly<FormContainerComponentProps>
+) {
+  return (
+    <NavigationDrawer      >
+      <FormContents {...props} />
+    </NavigationDrawer>
+  );
 }
 
 /**
@@ -41,35 +54,10 @@ interface FormContainerComponentProps {
  * @param {string} entityType The type of entity.
  * @param {FormType} formType The type of form such as add, update, delete, and view.
  * @param {boolean} isPrimaryEntity An optional indicator if the form is targeting a primary entity.
- * @param {boolean} isModal An optional indicator to render the form as a modal.
  */
-export default function FormContainerComponent(
+export function FormContainerComponent(
   props: Readonly<FormContainerComponentProps>
 ) {
-  if (props.isModal) {
-    const router = useRouter();
-    const currentEntityType: string = useSelector(getCurrentEntityType);
-    return (
-      <Drawer
-        onClose={() => {
-          if (currentEntityType != "") {
-            let redirectUrl: string = `${Routes.REGISTRY_GENERAL}/${currentEntityType}`;
-            if (currentEntityType === "outstanding") {
-              redirectUrl = Routes.REGISTRY_TASK_OUTSTANDING;
-            } else if (currentEntityType === "scheduled") {
-              redirectUrl = Routes.REGISTRY_TASK_SCHEDULED;
-            } else if (currentEntityType === "closed") {
-              redirectUrl = Routes.REGISTRY_TASK_CLOSED;
-            }
-            router.push(redirectUrl);
-          }
-        }}
-      >
-        <FormContents {...props} />
-      </Drawer>
-    );
-  }
-
   return (
     <div className=" flex flex-col w-full h-full mt-0  xl:w-[50vw] xl:h-[85vh] mx-auto justify-between py-4 px-4 md:px-8 bg-muted xl:border-1 xl:shadow-lg xl:border-border xl:rounded-xl xl:mt-4  ">
       <FormContents {...props} />
@@ -237,9 +225,7 @@ function FormContents(props: Readonly<FormContainerComponentProps>) {
   return (
     <>
       <section
-        className={`flex justify-between items-center text-nowrap text-foreground p-1 ${
-          !props.isModal ? "mt-0" : "mt-10"
-        }  mb-0.5  shrink-0`}
+        className={`flex justify-between items-center text-nowrap text-foreground p-1 mt-5 mb-0.5  shrink-0`}
       >
         <h1 className="text-xl font-bold">{`${translateFormType(
           props.formType,